feat: 라우터, 패키지 경로 재설정 및 레이아웃 디자인 변경
