+++ src/components/layouts/Footer.vue
@@ -3,72 +3,78 @@ import { useRoute } from 'vue-router'
 import IconGoal from '@/components/icons/IconGoal.vue'
 import IconMain from '@/components/icons/IconMain.vue'
 import IconSearch from '@/components/icons/IconSearch.vue'
+import IconRecommend from '@/components/icons/IconRecommend.vue'
 
 const route = useRoute()
-
-// 현재 라우터와 비교해서 활성화 여부 확인
 const isActive = (tabName) => route.name === tabName
 </script>
 
 <template>
   <div class="footer">
-    <router-link :to="{ name: 'policySearchPage' }">
-      <!-- <IconSearch :active="isActive"{} -->
-      <i class="fas fa-feather-alt"></i>
-      <div>탐색</div>
+    <router-link :to="{ name: 'searchPage' }" class="tab">
+      <IconSearch :active="isActive('searchPage')" />
+      <span :class="{ active: isActive('searchPage') }">탐색</span>
     </router-link>
 
-    <router-link :to="{ name: 'financialSearchPage' }">
-      <i class="fas fa-wave-square"></i>
-      <div>추천</div>
+    <router-link :to="{ name: 'recommendPage' }" class="tab">
+      <IconRecommend :active="isActive('recommendPage')" />
+      <span :class="{ active: isActive('recommendPage') }">추천</span>
     </router-link>
 
     <router-link :to="{ name: 'mainPage' }" class="tab">
       <IconMain :active="isActive('mainPage')" />
-      <div :class="{ active: isActive('mainPage') }">메인페이지</div>
+      <span :class="{ active: isActive('mainPage') }">메인</span>
     </router-link>
 
     <router-link :to="{ name: 'goalMain' }" class="tab">
       <IconGoal :active="isActive('goalMain')" />
-      <div :class="{ active: isActive('goalMain') }">목표</div>
+      <span :class="{ active: isActive('goalMain') }">목표</span>
     </router-link>
 
-    <router-link :to="{ name: 'scrapPage' }">
-      <i class="fas fa-tag"></i>
-      <div>스크랩</div>
+    <router-link :to="{ name: 'scrapPage' }" class="tab">
+      <i class="fas fa-tag" :style="{ color: isActive('scrapPage') ? '#1A1A1A' : '#C2C6CE' }"></i>
+      <span :class="{ active: isActive('scrapPage') }">자산</span>
+      <!-- 자산 컴포넌트로 변경할 부분 -->
     </router-link>
   </div>
 </template>
 
 <style scoped>
 .footer {
-  padding: 10px;
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  height: 60px;
+  padding: 8px 0;
   display: flex;
-  justify-content: space-between;
+  justify-content: space-around;
   align-items: center;
-  background: white;
+  background: white; /* 개별 컴포넌트에서 배경 관리 */
   border-top: 1px solid #eee;
+  box-shadow: 0 -1px 4px rgba(0, 0, 0, 0.05);
+  z-index: 1000;
 }
 
 .footer > .tab {
-  color: inherit;
-  font-size: 11pt;
   display: flex;
   flex-direction: column;
   align-items: center;
+  text-align: center;
   text-decoration: none;
+  color: inherit;
+  gap: 2px;
 }
 
-.footer .active {
+.footer .tab span {
+  font-size: 11px;
+  color: #C2C6CE;
+  line-height: 1.2;
+  font-weight: 400;
+}
+
+.footer .tab span.active {
   color: #1A1A1A;
   font-weight: 600;
-}
-
-.footer .tab div {
-  color: #C2C6CE;
-}
-
-.footer .tab div.active {
-  color: #1A1A1A;
 }
 </style>
