--- src/pages/policySearch/RegionSelectModal.vue
@@ -1,548 +1,0 @@
-<template>
-  <div class="modal-overlay" @click.self="$emit('close')">
-    <div
-      class="modal-content shadow rounded bg-white p-4"
-      style="max-width: 700px; width: 95%"
-    >
-      <!-- 상단 -->
-      <div class="d-flex justify-content-between align-items-center mb-3">
-        <h5 class="fw-bold mb-0">지역 선택</h5>
-        <i
-          class="fa-solid fa-xmark"
-          style="cursor: pointer"
-          @click="$emit('close')"
-        ></i>
-      </div>
-
-      <!-- 시도 + 시군구 -->
-      <div class="d-flex" style="gap: 1rem; height: 400px">
-        <!-- 시도 리스트 -->
-        <div
-          class="pe-3 border-end"
-          style="width: 60%; overflow-y: auto; height: 100%"
-        >
-          <div
-            v-for="sido in Object.keys(regionMap)"
-            :key="sido"
-            class="py-1 px-2"
-            :class="selectedSido === sido ? 'fw-bold bg-light rounded' : ''"
-            @click="selectSido(sido)"
-            style="cursor: pointer"
-          >
-            {{ sido }}
-            <span class="text-primary">{{ selectedCountBySido(sido) }}</span>
-          </div>
-        </div>
-
-        <!-- 시군구 리스트 -->
-        <div class="ps-2" style="width: 70%; overflow-y: auto; height: 100%">
-          <div class="form-check mb-2">
-            <input
-              class="form-check-input"
-              type="checkbox"
-              id="checkAll"
-              :checked="isAllChecked"
-              @change="toggleAllGugun"
-            />
-            <label class="form-check-label fw-bold" for="checkAll">전체</label>
-          </div>
-
-          <div
-            v-for="gugun in regionMap[selectedSido]"
-            :key="gugun"
-            class="form-check"
-          >
-            <input
-              class="form-check-input"
-              type="checkbox"
-              :id="gugun"
-              :value="gugun"
-              v-model="selectedGugun"
-            />
-            <label class="form-check-label" :for="gugun">{{ gugun }}</label>
-          </div>
-        </div>
-      </div>
-
-      <!-- 선택된 지역 -->
-      <div class="mt-3 selected-area-scroll">
-        <span
-          v-for="area in selectedFullNames"
-          :key="area"
-          class="badge bg-secondary me-1 mb-1"
-        >
-          {{ area }}
-          <i
-            class="fa-solid fa-xmark ms-1"
-            style="cursor: pointer"
-            @click="removeArea(area)"
-          ></i>
-        </span>
-      </div>
-
-      <!-- 하단 버튼 -->
-      <div class="d-flex justify-content-end gap-2 mt-4">
-        <button
-          class="btn btn-sm btn-outline-secondary"
-          @click="$emit('close')"
-        >
-          닫기
-        </button>
-        <button class="btn btn-sm btn-outline-secondary" @click="resetAll">
-          초기화
-        </button>
-        <button class="btn btn-sm btn-primary" @click="apply">적용하기</button>
-      </div>
-    </div>
-  </div>
-</template>
-
-<script setup>
-import { ref, reactive, computed, watch, onMounted } from 'vue';
-import { fetchZipCodes } from '@/api/regionApi';
-
-const emit = defineEmits(['close', 'apply']);
-
-// 지역 데이터
-const regionMap = {
-  전국: [],
-  서울특별시: [
-    '강남구',
-    '강동구',
-    '강북구',
-    '강서구',
-    '관악구',
-    '광진구',
-    '구로구',
-    '금천구',
-    '노원구',
-    '도봉구',
-    '동대문구',
-    '동작구',
-    '마포구',
-    '서대문구',
-    '서초구',
-    '성동구',
-    '성북구',
-    '송파구',
-    '양천구',
-    '영등포구',
-    '용산구',
-    '은평구',
-    '종로구',
-    '중구',
-    '중랑구',
-  ],
-
-  부산광역시: [
-    '강서구',
-    '금정구',
-    '기장군',
-    '남구',
-    '동구',
-    '동래구',
-    '부산진구',
-    '북구',
-    '사상구',
-    '사하구',
-    '서구',
-    '수영구',
-    '연제구',
-    '영도구',
-    '중구',
-    '해운대구',
-  ],
-
-  대구광역시: [
-    '군위군',
-    '남구',
-    '달서구',
-    '달성군',
-    '동구',
-    '북구',
-    '서구',
-    '수성구',
-    '중구',
-  ],
-
-  인천광역시: [
-    '강화군',
-    '계양구',
-    '남동구',
-    '동구',
-    '미추홀구',
-    '부평구',
-    '서구',
-    '연수구',
-    '옹진군',
-    '중구',
-  ],
-
-  광주광역시: ['광산구', '남구', '동구', '북구', '서구'],
-
-  대전광역시: ['대덕구', '동구', '서구', '유성구', '중구'],
-
-  울산광역시: ['남구', '동구', '북구', '울주군', '중구'],
-
-  세종특별자치시: ['세종특별자치시'],
-
-  경기도: [
-    '가평군',
-    '고양시 덕양구',
-    '고양시 일산동구',
-    '고양시 일산서구',
-    '과천시',
-    '광명시',
-    '광주시',
-    '구리시',
-    '군포시',
-    '김포시',
-    '남양주시',
-    '동두천시',
-    '부천시 소사구',
-    '부천시 오정구',
-    '부천시 원미구',
-    '성남시 분당구',
-    '성남시 수정구',
-    '성남시 중원구',
-    '수원시 권선구',
-    '수원시 영통구',
-    '수원시 장안구',
-    '수원시 팔달구',
-    '시흥시',
-    '안산시 단원구',
-    '안산시 상록구',
-    '안성시',
-    '안양시 동안구',
-    '안양시 만안구',
-    '양주시',
-    '양평군',
-    '여주시',
-    '연천군',
-    '오산시',
-    '용인시 기흥구',
-    '용인시 수지구',
-    '용인시 처인구',
-    '의왕시',
-    '의정부시',
-    '이천시',
-    '파주시',
-    '평택시',
-    '포천시',
-    '하남시',
-    '화성시',
-  ],
-  충청북도: [
-    '괴산군',
-    '단양군',
-    '보은군',
-    '영동군',
-    '옥천군',
-    '음성군',
-    '제천시',
-    '증평군',
-    '진천군',
-    '청주시 상당구',
-    '청주시 서원구',
-    '청주시 청원구',
-    '청주시 흥덕구',
-    '충주시',
-  ],
-
-  충청남도: [
-    '계룡시',
-    '공주시',
-    '금산군',
-    '논산시',
-    '당진시',
-    '보령시',
-    '부여군',
-    '서산시',
-    '서천군',
-    '아산시',
-    '예산군',
-    '천안시 동남구',
-    '천안시 서북구',
-    '청양군',
-    '태안군',
-    '홍성군',
-  ],
-
-  전라남도: [
-    '강진군',
-    '고흥군',
-    '곡성군',
-    '광양시',
-    '구례군',
-    '나주시',
-    '담양군',
-    '목포시',
-    '무안군',
-    '보성군',
-    '순천시',
-    '신안군',
-    '여수시',
-    '영광군',
-    '영암군',
-    '완도군',
-    '장성군',
-    '장흥군',
-    '진도군',
-    '함평군',
-    '해남군',
-    '화순군',
-  ],
-
-  경상북도: [
-    '경산시',
-    '경주시',
-    '고령군',
-    '구미시',
-    '김천시',
-    '문경시',
-    '봉화군',
-    '상주시',
-    '성주군',
-    '안동시',
-    '영덕군',
-    '영양군',
-    '영주시',
-    '영천시',
-    '예천군',
-    '울릉군',
-    '울진군',
-    '의성군',
-    '청도군',
-    '청송군',
-    '칠곡군',
-    '포항시 남구',
-    '포항시 북구',
-  ],
-
-  경상남도: [
-    '거제시',
-    '거창군',
-    '고성군',
-    '김해시',
-    '남해군',
-    '밀양시',
-    '사천시',
-    '산청군',
-    '양산시',
-    '의령군',
-    '진주시',
-    '창녕군',
-    '창원시 마산합포구',
-    '창원시 마산회원구',
-    '창원시 성산구',
-    '창원시 의창구',
-    '창원시 진해구',
-    '통영시',
-    '하동군',
-    '함안군',
-    '함양군',
-    '합천군',
-  ],
-
-  제주특별자치도: ['서귀포시', '제주시'],
-
-  강원특별자치도: [
-    '강릉시',
-    '고성군',
-    '동해시',
-    '삼척시',
-    '속초시',
-    '양구군',
-    '양양군',
-    '영월군',
-    '원주시',
-    '인제군',
-    '정선군',
-    '철원군',
-    '춘천시',
-    '태백시',
-    '평창군',
-    '홍천군',
-    '화천군',
-    '횡성군',
-  ],
-
-  전북특별자치도: [
-    '고창군',
-    '군산시',
-    '김제시',
-    '남원시',
-    '무주군',
-    '부안군',
-    '순창군',
-    '완주군',
-    '익산시',
-    '임실군',
-    '장수군',
-    '전주시 덕진구',
-    '전주시 완산구',
-    '정읍시',
-    '진안군',
-  ],
-};
-
-// 상태 관리
-const selectedSido = ref(Object.keys(regionMap)[0]);
-const selectedGugun = ref([]);
-const selectedGugunBySido = reactive({});
-
-// 컴포넌트가 로드될 때 전체 선택되지 않도록 보장
-onMounted(() => {
-  Object.keys(regionMap).forEach((sido) => {
-    selectedGugunBySido[sido] = []; // 모두 빈 배열로 초기화
-  });
-});
-
-const selectSido = (sido) => {
-  selectedSido.value = sido;
-  if (!selectedGugunBySido[sido]) {
-    selectedGugunBySido[sido] = [];
-  }
-  selectedGugun.value = selectedGugunBySido[sido];
-};
-
-// 구 선택
-watch(selectedGugun, (newVal) => {
-  selectedGugunBySido[selectedSido.value] = newVal;
-});
-
-// 전체 선택 토글
-const isAllChecked = computed(() => {
-  const all = regionMap[selectedSido.value] || [];
-
-  if (selectedSido.value === '전국') {
-    return false;
-  }
-
-  return selectedGugun.value.length === all.length;
-});
-
-const toggleAllGugun = (e) => {
-  const checked = e.target.checked;
-
-  // 전국 전체 선택
-  if (selectedSido.value === '전국') {
-    Object.entries(regionMap).forEach(([sido, guguns]) => {
-      if (!selectedGugunBySido[sido]) selectedGugunBySido[sido] = [];
-      selectedGugunBySido[sido] = checked ? [...guguns] : [];
-    });
-  } else {
-    // 기존 시도별 전체 선택
-    const all = regionMap[selectedSido.value];
-    selectedGugunBySido[selectedSido.value] = checked ? [...all] : [];
-  }
-
-  // 현재 선택중인 시도 반영
-  selectedGugun.value = selectedGugunBySido[selectedSido.value];
-};
-
-// 선택된 구 개수 표시
-const selectedCountBySido = (sido) => {
-  return selectedGugunBySido[sido]?.length || 0;
-};
-
-// 지역 전체 이름
-const selectedFullNames = computed(() =>
-  Object.entries(selectedGugunBySido).flatMap(([sido, guguns]) =>
-    guguns.map((g) => `${sido} ${g}`)
-  )
-);
-
-// 개별 구 제거
-const removeArea = (fullName) => {
-  const [sido, gugun] = fullName.split(' ');
-  selectedGugunBySido[sido] = selectedGugunBySido[sido].filter(
-    (g) => g !== gugun
-  );
-  if (selectedSido.value === sido) {
-    selectedGugun.value = selectedGugunBySido[sido];
-  }
-};
-
-// 초기화
-const resetAll = () => {
-  Object.keys(selectedGugunBySido).forEach((sido) => {
-    selectedGugunBySido[sido] = [];
-  });
-  selectedGugun.value = [];
-};
-const apply = async () => {
-  try {
-    const fullNames = selectedFullNames.value; //
-    const zipCodes = await fetchZipCodes(fullNames);
-
-    emit('apply', { zipCodes, regionNames: fullNames }); // zipCodes + 지역명 함께 전달
-    emit('close');
-  } catch (err) {
-    console.error('지역 코드 가져오기 실패:', err);
-    alert('지역 필터 적용 중 오류가 발생했습니다.');
-  }
-};
-</script>
-
-<style>
-.modal-overlay {
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100%;
-  height: 100%;
-  background-color: rgba(0, 0, 0, 0.4);
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  z-index: 1000;
-}
-
-.modal-content {
-  background-color: white;
-  border-radius: 8px;
-  width: 100%;
-  max-width: 392px;
-  max-height: 90vh;
-  overflow-y: auto;
-  padding: 1rem;
-  box-sizing: border-box;
-  font-size: 0.8rem; /* ✅ 전체 폰트 축소 */
-  display: flex;
-  flex-direction: column; /* ✅ 아래 버튼까지 포함한 수직 정렬 */
-}
-
-.selected-area-scroll {
-  max-height: 120px; /* ✅ 선택된 지역 badge 영역 고정 */
-  overflow-y: auto;
-  border: 1px solid #ddd;
-  padding: 0.5rem;
-  border-radius: 4px;
-  background-color: #f8f9fa;
-}
-
-/* 왼쪽 시도 리스트 스크롤바 */
-.pe-3::-webkit-scrollbar {
-  width: 4px;
-}
-.pe-3::-webkit-scrollbar-thumb {
-  background-color: #ccc;
-  border-radius: 3px;
-}
-.pe-3::-webkit-scrollbar-track {
-  background-color: transparent;
-}
-
-/* 오른쪽 시군구 리스트 스크롤바 */
-.ps-2::-webkit-scrollbar {
-  width: 4px;
-}
-.ps-2::-webkit-scrollbar-thumb {
-  background-color: #ccc;
-  border-radius: 3px;
-}
-.ps-2::-webkit-scrollbar-track {
-  background-color: transparent;
-}
-</style>
