+++ src/pages/search/financialSearch/util/bankLogo.js
