+++ src/router/policySearch.js
@@ -3,13 +3,13 @@ export default [
   {
     path: '/policySearch/PolicySearchPage',
     name: 'policySearchPage',
-    component: () => import('../pages/policySearch/PolicySearchPage.vue'),
+    component: () => import('../pages/search/policySearch/PolicySearchPage.vue'),
     meta: { isHeader: true, isFooter: true },
   },
   {
     path: '/policySearch/detail/:id',
     name: 'policyDetail',
-    component: () => import('../pages/policySearch/policyDetailPage.vue'),
+    component: () => import('../pages/search/policySearch/policyDetailPage.vue'),
     meta: { isHeader: true, isFooter: true },
   },
 ];
