+++ src/router/recommend.js
@@ -1,0 +1,8 @@
+export default [
+    {
+      path: '/recommend',
+      name: 'recommendPage',
+      component: () => import('../pages/recommend/RecommendPage.vue'),
+      meta: { isHeader: true, isFooter: true },
+    }
+  ];
