+++ src/router/search.js
@@ -1,0 +1,8 @@
+export default [
+    {
+      path: '/search',
+      name: 'searchPage',
+      component: () => import('../pages/search/SearchGuidePage.vue'),
+      meta: { isHeader: true, isFooter: true },
+    }
+  ];
